Synchronize IncidentRecord access
I realized that, since the same IncidentRecord instance is shared between
request handler threads, access to its state needs to be synchronized.
This is accomplished via the Groovy @Synchronized AST transformation.

However, because Ratpack is an asynchronous framework, we shouldn't
allow the request handling threads to block while waiting to access the
IncidentRecord state. To prevent this, access to the IncidentRecord now
occurs through Ratpack's Blocking facility, which defers the actual
IncidentRecord access to another thread pool.